// src/features/tracks/api/tracksApi.ts
import { baseApi } from '@/app/api/baseApi.ts'
import type { FetchTracksResponse } from './trackApi.types.ts'
import { fetchTracksResponseSchema } from '@/features/tracks/model/tracks.schemas.ts'
import { withZodCatch } from '@/common/utils'

export const tracksApi = baseApi.injectEndpoints({
    endpoints: build => ({
        fetchTracks: build.infiniteQuery<
            FetchTracksResponse,
            void,
            string | null
        >({
            infiniteQueryOptions: {
                initialPageParam: null,
                getNextPageParam: lastPage => {
                    return lastPage.meta.nextCursor || null
                },
            },
            query: ({ pageParam }) => ({
                url: 'playlists/tracks',
                params: {
                    cursor: pageParam,
                    paginationType: 'cursor',
                    pageSize: 10,
                },
            }),
            ...withZodCatch(fetchTracksResponseSchema),
        }),
    }),
})

export const { useFetchTracksInfiniteQuery } = tracksApi
